修复了在极端情况下攻击动画不播放而导致unit卡死
重写了Actor2的triggerAnimation里的改变方向的逻辑。

+++ Client/Units/char_101_sora_factory.h
@@ -37,7 +37,7 @@ protected:
 	void setAnimationInfoStorage();
 
 	std::unique_ptr<Model::IAnimationPose> m_pose;
-	Model::AnimationInfo m_info[static_cast<size_t>(Game::IActor::AnimationStatus::COUNT)];
+	Model::AnimationInfo m_info[Game::IActor::AnimationStatus::COUNT];
 };
 
 }
