+++ Client/Units/char_151_myrtle_factory.cpp
@@ -68,7 +68,7 @@ void Char_151_Mytle_Factory::setAnimationInfoStorage() {
 	Model::IAnimationPose* pose;
 	pose = m_pose[0].get();
 	{
-		info = m_info[0] + (size_t)Game::IActor::AnimationStatus::Normal;
+		info = m_info[0] + Game::IActor::AnimationStatus::Normal;
 		{
 			info->Default = (Model::AnimationInfo::InfoType)pose->getAnimation("Default");
 			info->Begin = (Model::AnimationInfo::InfoType)pose->getAnimation("Start");
@@ -77,9 +77,9 @@ void Char_151_Mytle_Factory::setAnimationInfoStorage() {
 			info->Die = (Model::AnimationInfo::InfoType)pose->getAnimation("Die");
 			info->StunIn = info->Die;
 		}
-		m_info[0][(size_t)Game::IActor::AnimationStatus::Skill0] =
-			m_info[0][(size_t)Game::IActor::AnimationStatus::Normal];
-		info = m_info[0] + (size_t)Game::IActor::AnimationStatus::Skill0;
+		m_info[0][Game::IActor::AnimationStatus::Skill0] =
+			m_info[0][Game::IActor::AnimationStatus::Normal];
+		info = m_info[0] + Game::IActor::AnimationStatus::Skill0;
 		{
 			info->Begin = (Model::AnimationInfo::InfoType)pose->getAnimation("Skill_Begin");
 			info->Idle = (Model::AnimationInfo::InfoType)pose->getAnimation("Skill_Loop");
@@ -88,7 +88,7 @@ void Char_151_Mytle_Factory::setAnimationInfoStorage() {
 	}
 	pose = m_pose[1].get();
 	{
-		info = m_info[1] + (size_t)Game::IActor::AnimationStatus::Normal;
+		info = m_info[1] + Game::IActor::AnimationStatus::Normal;
 		{
 			info->Default = (Model::AnimationInfo::InfoType)pose->getAnimation("Default");
 			info->Begin = (Model::AnimationInfo::InfoType)pose->getAnimation("Start");
@@ -97,9 +97,9 @@ void Char_151_Mytle_Factory::setAnimationInfoStorage() {
 			info->Die = (Model::AnimationInfo::InfoType)pose->getAnimation("Die");
 			info->StunIn = info->Die;
 		}
-		m_info[1][(size_t)Game::IActor::AnimationStatus::Skill0] =
-			m_info[1][(size_t)Game::IActor::AnimationStatus::Normal];
-		info = m_info[1] + (size_t)Game::IActor::AnimationStatus::Skill0;
+		m_info[1][Game::IActor::AnimationStatus::Skill0] =
+			m_info[1][Game::IActor::AnimationStatus::Normal];
+		info = m_info[1] + Game::IActor::AnimationStatus::Skill0;
 		{
 			info->Begin = (Model::AnimationInfo::InfoType)pose->getAnimation("Skill_Begin");
 			info->Idle = (Model::AnimationInfo::InfoType)pose->getAnimation("Skill_Loop");
