+++ Client/Units/char_151_myrtle_factory.h
@@ -37,7 +37,7 @@ protected:
 	void setAnimationInfoStorage();
 
 	std::unique_ptr<Model::IAnimationPose> m_pose[2];
-	Model::AnimationInfo m_info[2][static_cast<size_t>(Game::IActor::AnimationStatus::COUNT)];
+	Model::AnimationInfo m_info[2][Game::IActor::AnimationStatus::COUNT];
 };
 
 }
